Spread out the views to their respective view files to make them more modular and breathable. Also created a package called views to contain these files, and edited the urls.py to make for these changes
