--- django_project/account/templates/recipient/recipient_dashboard.html
@@ -1,65 +1,0 @@
-{% extends 'dashboard_base.html' %}
-
-{% load static %}
-
-
-
-{% block profile_image %} 
-    <img src="{% static '/img/team/team-2.jpg' %}" class="rounded mx-auto d-block" alt="..." width="80px" height="80px" >
-{% endblock %}
-
-
-
-{% block content %}
-
-      <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
-        <h1 class="h2">{{ greeting_message }}, {{ user.username }}</h1>
-        <div class="btn-toolbar mb-2 mb-md-0">
-          <div class="btn-group me-2">
-            <!-- <button type="button" class="btn btn-sm btn-outline-secondary">Share</button> -->
-            <button type="button" class="btn btn-sm btn-outline-secondary">notifications</button>
-          </div>
-          <!-- <button type="button" class="btn btn-sm btn-outline-secondary dropdown-toggle">
-            <span data-feather="calendar"></span>
-            This week
-          </button> -->
-        </div>
-      </div>
-
-
-      <h2>Packages</h2>
-      <div class="table-responsive">
-        <table class="table table-striped table-sm">
-          <thead>
-            <tr>
-              <th scope="col">No.</th>
-              <th scope="col">Package Title</th>
-              <th scope="col">Devilering To</th>
-              <th scope="col">Status</th>
-            </tr>
-          </thead>
-          <tbody>
-            <!-- <tr>
-                <td>01</td>
-                <td>Fridge</td>
-                <td>MAQULA</td>
-                <td><button class="btn btn-sm btn-primary text-white w-50">upcoming</button></td>
-              </tr>
-              <tr>
-                <td>02</td>
-                <td>Sofa Set</td>
-                <td>MAQULA</td>
-                <td><button class="btn btn-sm btn-info text-white w-50">Ongoing</button></td>
-              </tr>
-              <tr>
-                <td>03</td>
-                <td>Five Bags of Irish</td>
-                <td>Total Wandegeya</td>
-                <td><button class="btn btn-sm btn-success text-white w-50">Completed</button></td>
-              </tr> -->
-              
-            
-          </tbody>
-        </table>
-      </div>
-   {% endblock %}
