+++ django_project/account/urls.py
@@ -1,28 +1,45 @@
 from django.urls import path
-from . import views
 from django.contrib.auth import views as auth_views
+from .views import (
+    site_views,
+    sender_views,
+    courier_views,
+    admin_views,
+    authentication_views
+)
 
 urlpatterns = [
 
-    path('', views.index, name='home'),
-    path('base/', views.base, name='base'),
-    path('services/', views.services, name='services'),
-    path('about/', views.about, name='about'),
-    path('contact/', views.contact, name='contact'),
-    path('login/', views.login_view, name='login_view'),
-    path('register/', views.register, name='register'),
+    # site routes
+    path('', site_views.index, name='home'),
+    path('services/', site_views.services, name='services'),
+    path('about/', site_views.about, name='about'),
+    path('contact/', site_views.contact, name='contact'),
+
+
+    # authentication routes
+    path('login/', authentication_views.login_view, name='login_view'),
+    path('register/', authentication_views.register, name='register'),
     path('logout/', auth_views.LogoutView.as_view(template_name='index.html'), name='logout_user'),
-    path('admin_dashboard/', views.admin, name='admin_dashboard'),
-    path('sender_dashboard/', views.sender_dashboard, name='sender_dashboard'),
-    path('register_package/', views.register_package, name='register_package'),
-    path('recipient_dashboard/', views.recipient_dashboard, name='recipient_dashboard'),
-    path('courier_dashboard/', views.courier_dashboard, name='courier_dashboard'),
-    path('users/', views.users, name='users'),
-    path('riders/', views.riders, name='riders'),
-    path('admin_history/', views.admin_history, name='admin_history'),
-    path('assign_courier/<int:package_id>/', views.assign_courier, name='assign_courier'),
-    path('notify_arrival/<int:package_id>/', views.notify_arrival, name='notify_arrival'),
-    path('confirm_delivery/<int:package_id>/',views.confirm_delivery, name='confirm_delivery'),
-    path('sender_history/', views.sender_history, name='sender_history'),
-    path('courier_history/', views.courier_history, name='courier_history'),
+
+
+    # admin routes
+    path('admin_dashboard/', admin_views.admin, name='admin_dashboard'),
+    path('users/', admin_views.users, name='users'),
+    path('riders/', admin_views.riders, name='riders'),
+    path('admin_history/', admin_views.admin_history, name='admin_history'),
+    path('assign_courier/<int:package_id>/', admin_views.assign_courier, name='assign_courier'),
+
+
+    # sender routes
+    path('sender_dashboard/', sender_views.sender_dashboard, name='sender_dashboard'),
+    path('register_package/', sender_views.register_package, name='register_package'),
+    path('sender_history/', sender_views.sender_history, name='sender_history'),
+
+
+    # courier routes
+    path('courier_dashboard/', courier_views.courier_dashboard, name='courier_dashboard'),
+    path('notify_arrival/<int:package_id>/', courier_views.notify_arrival, name='notify_arrival'),
+    path('confirm_delivery/<int:package_id>/', courier_views.confirm_delivery, name='confirm_delivery'),
+    path('courier_history/', courier_views.courier_history, name='courier_history'),
 ]
