+++ django_project/account/views/site_views.py
@@ -1,0 +1,14 @@
+from django.shortcuts import render
+
+
+def index(request):
+    return render(request, 'index.html', {})
+
+def about(request):
+    return render(request, 'about.html', {})
+
+def services(request):
+    return render(request, 'services.html', {})
+
+def contact(request):
+    return render(request, 'contact.html', {})
